feat: change student id type to integer

+++ migrations/init_db.sql
@@ -1,7 +1,7 @@
 CREATE TABLE IF NOT EXISTS student
 (
-    student_id      BIGINT          NOT NULL,
-    role            VARCHAR(30),
+    student_id INTEGER NOT NULL,
+    role       VARCHAR(30),
 
     UNIQUE (student_id),
     PRIMARY KEY (student_id)
@@ -9,11 +9,11 @@ CREATE TABLE IF NOT EXISTS student
 
 CREATE TABLE IF NOT EXISTS employee
 (
-    employee_id         BIGINT          GENERATED ALWAYS AS IDENTITY,
-    employee_name       VARCHAR(255)    NOT NULL,
-    description         TEXT            NOT NULL,
-    phone               VARCHAR(255)    NOT NULL,
-    email               VARCHAR(255)    NOT NULL,
+    employee_id   BIGINT GENERATED ALWAYS AS IDENTITY,
+    employee_name VARCHAR(255) NOT NULL,
+    description   TEXT         NOT NULL,
+    phone         VARCHAR(255) NOT NULL,
+    email         VARCHAR(255) NOT NULL,
 
-    PRIMARY KEY(employee_id)
+    PRIMARY KEY (employee_id)
 );
